Mast limit debug attempt - 1
Bugs: When mast is at its angle limit, strange behavior can be noted. 
No.1 (Intermittent) is that in manual mode when a board sw is still pressed when the mast is a limit, the mcc begins pulsating until it gets to the point that the angle value wraps around (at -180). It then shift to the positive limit (180). 
No2 (Can be reproduced w consistency) is that when the mast stops correctly at its  angle limit and that the sw for the opposite direction is pressed, the motor begins rotating non stop in the ordered direction. The problem is that even if you press shortly on the sw, the mast acts as if your finger is still on the sw ;the user must press the button a second time to make it stop.

Debug steps taken: 
No1: proper sw management has been checked by assignating debug leds to their activation and releases. 
No2: StateManual function has been checked by assingating leds to when an angle limit is reached and left (but is not really functionning properly_
No3: The slowstop