Mast limit fix #1
The problem where the motor was pulsating when at a MAX or MIN limit is
solved.
The problem was in fact in the StateGetMastData(). The limit test was
made so that the MastManualStop function was being called repeatedly.
The bug where the mast doesnt stop at a single button's push after a
limit stop has yet to be fixed.